Fix bug where sometimes dependencies were not loaded correctly
For example when first k is loaded and executed. n is in the same file
as k. Then n is loaded, which has a dependency to o, which is in a
second file. Even if n is loaded, it should still check and load its
dependencies, so the second file will be loaded.

Write a test with mapping configuration which exposed the problem.

Split the npm scripts commands into separate scripts.