Fix #9 - Label modules as how they are originally required.
Fix example if "foo" is in the map, which could be defined in the
browser package.json field, or could be a node_module package, label
that module like "foo", and not "./node_modules/foo". Similarly if a
module is required as "./bar", which refers to "bar/index.js", label it
as "./bar" and not "./bar/index".